requirements.txt
Flask==2.3.2
torch==2.2.0
Pillow==10.0.0
torchvision==0.17.0
fastapi==0.111.0
uvicorn==0.19.0
pydantic==2.7.2

